requirements.txt
annotated-types==0.7.0
emoji==2.12.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
pydantic==1.10.15
pydantic-sqlalchemy==0.0.9
pydantic_core==2.18.3
smmap==5.0.1
SQLAlchemy==1.4.52
tabulate==0.9.0
typing_extensions==4.12.0
openai==1.33.0
python-dotenv==1.0.0